set up member and completion express routers
